More progress on const generics support

+++ src/kern/kernel_nm.rs
@@ -69,7 +69,7 @@ impl<T: Scalar, At: Mat<T>, Bt: Mat<T>, Ct: Mat<T>, const Nr: usize, const Mr: u
                     prefetch_c_row(next_c_ir.offset(3*c_leaf_rs)); 
                 }
 
-                if Ct::full_leaves() || (n - jr >= Nr) && (m - ir >= Mr) {
+                if Ct::full_leaves() || (n - jr >= Nr as isize) && (m - ir >= Mr as isize) {
                     <UkernelWrapper<T,{Mr},{Nr}>>::run(k, &mut alpha, a_ir, b_jr, &mut beta, c_ir, c_leaf_rs, c_leaf_cs);
                 } else {
 					let tp = self.tmp.get_mut_buffer();
@@ -78,8 +78,8 @@ impl<T: Scalar, At: Mat<T>, Bt: Mat<T>, Ct: Mat<T>, const Nr: usize, const Mr: u
                     let t_cs = self.tmp.get_column_stride() as isize;
 					<UkernelWrapper<T,{Mr},{Nr}>>::run(k, &mut alpha, a_ir, b_jr, &mut t_scalar, tp, t_rs, t_cs);
                     
-                    let local_m = if m-ir >= Mr { Mr } else { m-ir };
-                    let local_n = if n-jr >= Nr { Nr } else { n-jr };
+                    let local_m = if m-ir >= Mr as isize { Mr as isize } else { m-ir };
+                    let local_n = if n-jr >= Nr as isize { Nr as isize } else { n-jr };
 
 					//Add t to c
                     for ii in 0..local_m {
@@ -91,11 +91,11 @@ impl<T: Scalar, At: Mat<T>, Bt: Mat<T>, Ct: Mat<T>, const Nr: usize, const Mr: u
                     }
                 }
 
-                ir += Mr;
+                ir += Mr as isize;
                 a_ir = a_ir.offset(a_mr_stride);
                 c_ir = c_ir.offset(c_mr_stride);
             }
-            jr += Nr;
+            jr += Nr as isize;
             c_jr = c_jr.offset(c_nr_stride);
             b_jr = b_jr.offset(b_nr_stride);
         }
